[RGTC-593] Backport Y Schedules
