+++ modules/custom/openy_repeat/js/repeat/.jshintrc
@@ -1,0 +1,3 @@
+{
+  "esnext": true
+}
