+++ modules/custom/openy_repeat/js/repeat/README.MD
@@ -1,0 +1,4 @@
+### Compile repeat.js
+- Run command `npm install` for install dependency
+- Run command `npm run build` for compile repeat.js to parent folder "src"
+- Run `drush cr` in docroot folder
