+++ modules/custom/openy_repeat/js/repeat/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "repeat.js",
+  "version": "1.0.0",
+  "description": "npm module for recompile repeat.js",
+  "main": "repeat.js",
+  "scripts": {
+    "build": "webpack --config webpack.config.js --mode=production"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "@babel/polyfill": "^7.6.0"
+  },
+  "devDependencies": {
+    "@babel/cli": "^7.6.2",
+    "@babel/core": "^7.6.2",
+    "@babel/preset-env": "^7.6.2",
+    "@vue/babel-preset-app": "^3.11.0",
+    "babel-loader": "^8.0.6",
+    "webpack": "^4.41.0",
+    "webpack-cli": "^3.3.9"
+  }
+}
